Fixed duration event not being fired when playing the same media url + Fixed double invocation of the ClosePlayerActivity when closed through the close button in the header

+++ application/src/main/java/com/github/yoep/popcorn/watched/WatchedService.java
@@ -101,7 +101,7 @@ public class WatchedService {
             if (time == ClosePlayerActivity.UNKNOWN || duration == ClosePlayerActivity.UNKNOWN)
                 return;
 
-            long percentageWatched = (time / duration) * 100;
+            double percentageWatched = ((double) time / duration) * 100;
             Media media = activity.getMedia();
 
             // check if the media has been watched for the percentage threshold
